removed image handler + made team detector nice
